+++ CHANGELOG.md
@@ -13,6 +13,10 @@ For instructions, see the [changelog confluence page](https://epcpower.atlassian
 - SC-590: Added access level column to the SunSpec and static modbus spreadsheet output.
 - SC-572: Added changelog for release notes.
 
+### Changed
+
+- SC-683: Changed rejected callback handling to use automatically generated interface
+
 ### Removed
 
 - SC-623: Removed rejected callback field from c interface structures.